add minimum coverage threshold of 1

+++ src/Options.cpp
@@ -22,7 +22,7 @@ extern float siteCovThreshold;
 extern double scoreThresh;
 extern double covSkew;
 extern bool all;
-extern double covThresh;
+extern unsigned covThresh;
 extern double maxCov;
 extern bool dupes;
 //extern uint64_t minSites;
